Code refact and more UT
